+++ src/vue/forms/AssetForm/AssetForm.spec.js
@@ -1,4 +1,4 @@
-import AssetForm from './index'
+import AssetForm from './AssetForm'
 
 import Vuelidate from 'vuelidate'